+++ src/main/java/johkar2/kth/se/moviedatabase/model/Movie.java
@@ -2,7 +2,7 @@ package johkar2.kth.se.moviedatabase.model;
 
 public class Movie extends Media {
 
-    public Movie(String title, String description, int rating){
+    public Movie(String title, String description, double rating){
         super(title,description,rating);
     }
 }
